+++ DiscordBot/BattleSystem/Entities/Weapons/Fist.cs
@@ -9,7 +9,7 @@ namespace DiscordBot.BattleSystem.Entities.Weapons
 {
     public class Fist : IWeapon
     {
-        public int Damage { get; set; }
+        public float Damage { get; set; }
         public string Name { get; set; }
         public Fist()
         {
